refactor(queries): extract duplicate selector_span into queries/mod.rs
`selector_span(selector: &MessageSelector) -> Option<Span>` was defined
identically in four language-service query modules: senders_query,
all_sends_query, announce_sites_query, and ffi_sites_query. Each copy
returned the merged span of the first and last keyword parts (or None
for unary/binary selectors) — nine lines of code repeated four times.

Extract the function once into `queries/mod.rs` as `pub(crate)`, and
replace the four private copies with `use super::selector_span;`.
