+++ crates/beamtalk-core/src/queries/announce_sites_query.rs
@@ -48,6 +48,7 @@
 //! Parse errors are tolerated: any sub-trees that parsed successfully still
 //! contribute results. A completely unparseable source returns an empty list.
 
+use super::selector_span;
 use crate::ast::{Expression, Pattern, StringSegment};
 use crate::source_analysis::{Span, lex_with_eof, parse};
 
@@ -336,17 +337,6 @@ fn selector_line(selector: &crate::ast::MessageSelector, fallback: Span, source:
     selector_span(selector)
         .unwrap_or(fallback)
         .line_number(source)
-}
-
-fn selector_span(selector: &crate::ast::MessageSelector) -> Option<Span> {
-    match selector {
-        crate::ast::MessageSelector::Keyword(parts) if !parts.is_empty() => {
-            let first = parts.first().unwrap().span;
-            let last = parts.last().unwrap().span;
-            Some(first.merge(last))
-        }
-        _ => None,
-    }
 }
 
 #[cfg(test)]
